turso y cambio en offer

+++ src/controllers/offer.controller.ts
@@ -1,75 +1,94 @@
 import { NextFunction, Request, Response } from "express"
 import { OfferService } from "../services/offer.service"
+import { HttpException } from "@/exceptions/httpException";
 
-export class OfferController {
-    static async getById(req: Request, res: Response, next: NextFunction) {
-        try {
+export class OfferController{
+    static async getById(req:Request, res:Response, next:NextFunction){
+        try{
             const id = Number.parseInt(req.params.id)
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
+            // pasar a entero
             const offer = await OfferService.getById(id)
             res.status(200).json(offer)
-        } catch (error) {
+        }catch(error){
             next(error)
         }
     }
 
-    static async getAll(req: Request, res: Response, next: NextFunction) {
-        try {
-            const user = await OfferService.getAll()
+    static async getAll(req:Request, res:Response, next: NextFunction){
+        try{
+            const { title } = req.query;
+            const user = await OfferService.getAll(title as string)
             res.status(200).json(user)
-        } catch (error) {
+        }catch(error){
             next(error)
         }
     }
 
-    static async create(req: Request, res: Response, next: NextFunction) {
-        try {
+    static async create(req:Request, res:Response, next: NextFunction){
+        try{
             const offerData = req.body
-            const newOffer = await OfferService.create(offerData)
+            const userId = req.body.user.id
+
+            if (!userId) throw new HttpException(400, "User creator ID is required");
+            
+
+            const newOffer = await OfferService.create(userId, offerData)
             res.status(200).json(newOffer)
-        } catch (error) {
+        }catch(error){
             next(error)
         }
     }
 
-    static async update(req: Request, res: Response, next: NextFunction) {
-        try {
+    static async update(req:Request, res:Response, next: NextFunction){
+        try{
             const offerData = req.body
             const id = Number.parseInt(req.params.id)
-            const updateOffer = await OfferService.update(id,offerData)
-            res.status(200).json(updateOffer)
-        } catch (error) {
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
+            const updatedOffer = await OfferService.update(id, offerData)
+            res.status(200).json(updatedOffer)
+        }catch(error){
             next(error)
         }
     }
 
-    static async delete(req: Request, res: Response, next: NextFunction) {
-        try {
+    static async delete(req:Request, res:Response, next: NextFunction){
+        try{
             const id = Number.parseInt(req.params.id)
-            const deleteOffer = await OfferService.delete(id)
-            res.status(200).json(deleteOffer)
-        } catch (error) {
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
+            const deletedOffer = await OfferService.delete(id)
+            res.status(200).json(deletedOffer)
+        }catch(error){
             next(error)
         }
     }
 
-    static async rate(req: Request, res: Response, next: NextFunction) {
-        try {
+    static async rate(req:Request, res:Response, next: NextFunction){
+        try{
             const id = Number.parseInt(req.params.id)
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
             const {value} = req.body
             const userId = req.body.user.id
+
             await OfferService.rate(userId, id, value)
-            res.status(200).json({message: 'offer rate succesfully'})
-        } catch (error) {
+            res.status(200).json({message: 'Offer rate successfully'})
+        }catch(error){
             next(error)
         }
     }
 
-    static async getRate(req: Request, res: Response, next: NextFunction) {
-        try {
+    static async getRate(req:Request, res:Response, next: NextFunction){
+        try{
             const id = Number.parseInt(req.params.id)
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
             await OfferService.getRate(id)
-            res.status(200).json({message: 'offer rate succesfully'})
-        } catch (error) {
+            res.status(200).json({message: 'Offer rate successfully'})
+        }catch(error){
             next(error)
         }
     }
